+++ indexer.php
@@ -9,7 +9,7 @@
 * @license GNU General Public License version 2 or later
 *     http://www.opensource.org/licenses/gpl-license.php
 *
-*  Copyright (C) 2018 by the following authors:
+*  Copyright (C) 2018-2019 by the following authors:
 *   Mark R. Evans   mark AT glfusion DOT org
 *
 *  Based on the Searcher Plugin for glFusion
@@ -25,7 +25,7 @@ global $_DB_table_prefix, $_TABLES;
 
 $_INDEXER_CONF['pi_name']            = 'indexer';
 $_INDEXER_CONF['pi_display_name']    = 'Indexer';
-$_INDEXER_CONF['pi_version']         = '1.0.0';
+$_INDEXER_CONF['pi_version']         = '1.0.1';
 $_INDEXER_CONF['gl_version']         = '1.7.6';
 $_INDEXER_CONF['pi_url']             = 'https://www.glfusion.org';