+++ language/english_utf-8.php
@@ -2,12 +2,14 @@
 /**
 * glFusion CMS
 *
-* English (UTF-8) Language File
+* Indexer - Indexer Plugin for glFusion
+*
+* Language File - UTF-8
 *
 * @license GNU General Public License version 2 or later
 *     http://www.opensource.org/licenses/gpl-license.php
 *
-*  Copyright (C) 2018 by the following authors:
+*  Copyright (C) 2018-2019 by the following authors:
 *   Mark R. Evans   mark AT glfusion DOT org
 *
 *  Based on the Searcher Plugin for glFusion
@@ -39,6 +41,7 @@ $LANG_INDEXER = array(
     'two_years'     => '2 years',
     'seconds'       => 'secs',
     'search_placeholder' => 'Enter Your Search Request',
+    'altered_search' => 'Using <b>all of these words</b> did not return any data so the search was modified to <b>any of these words</b>',
 );
 
 $LANG_INDEXER_ADM = array(
@@ -52,7 +55,7 @@ $LANG_INDEXER_ADM = array(
     'error_heading'     => 'Errors',
     'hlp_counters'      => 'Here are the search queries made by site visitors, along with the number of times each query was made.',
     'hlp_gen_all'       => 'Re-generate all indexes for the selected content types. Use this option after installing the plugin, or after changing certain key configuration items such as the minimum word length or changing the stemmer.',
-    'hlp_reindex'       => 'Reindexing content will remove all existing search items for the content type and re-scan the content to rebuild the search word index. This can take a significant amount of time on large volume content types such as Forums.',
+    'hlp_reindex'       => 'Re-indexing content will remove all existing search items for the content type and re-scan the content to rebuild the search word index. This can take a significant amount of time on large volume content types such as Forums.',
     'index_instructions'=> 'This will scan the selected content types and rebuild the Indexer index',
     'index_status'      => 'Indexing Status',
     'indexer_admin'     => 'Indexer Admin',
@@ -60,8 +63,8 @@ $LANG_INDEXER_ADM = array(
     'no_errors'         => 'No Errors',
     'overall_progress'  => 'Overall Progress',
     'queries'           => 'Queries',
-    'reindex_button'    => 'Reindex',
-    'reindex_title'     => 'Reindex Content',
+    'reindex_button'    => 'Re-index',
+    'reindex_title'     => 'Re-index Content',
     'remove_content_1'  => 'Removing existing index entries for ',
     'remove_content_2'  => ' - This can take several minutes....',
     'remove_fail'       => 'Failed to remove existing index entries.',
@@ -95,6 +98,12 @@ $LANG_fs['indexer'] = array(
     'fs_main' => 'Settings',
 );
 
+$LANG_configSelect['indexer'] = array(
+    0  => array(1 => 'True', 0 => 'False'),
+    1  => array('all' => 'All of the words', 'any' => 'Any of the words', 'phrase' => 'Exact Phrase'),
+);
+
+// glFusion v1.7.x support
 $LANG_configselects['indexer'] = array(
     0  => array('True' => 1, 'False' => 0 ),
     1  => array('All of the words' => 'all', 'Any of the words' => 'any', 'Exact Phrase' => 'phrase'),