+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Indexer Plugin ChangeLog
 
+## v1.0.1
+  - Resolved an issue where saving an article did not properly re-index the content
+  - PHP v7.4 fixes
+  - Search using any words if all words did not produce results
+  - Fixed an issue where some content saved would not update the search index
+
 ## v1.0.0
   - Add option for default search mode (all, any, phrase)
   - Added search time to show the number of seconds to perform search